Switch back to WADO-URI, rework VR routes, and other fixes
This includes adding an event that triggers when a volume
has completed loading, which allows the viewports to listen
for this and add the Masking segmentation only once the
volume has been fully loaded (this is what allows us to
switch back to WADO-URI).

This commit also adds z-decimation for very large volumes
(those with more than 2000 slices) which previously failed
to render.