Add a magic error trap for well-known error
This is hacky and should be only temporary, and in fact is
setup so that it will only run in development. This captures
"well-known" errors (currently just the one thrown from updateSurfaceData.js
when updating the 3d surface sometimes) and prevents them from ever
reaching the Webpack Dev server error overlay. They are instead printed
to the console as a warning (complete with stack trace).